factor out event sourced projection testcases
