+++ eventsourcing/tests/projection.py
@@ -6,19 +6,23 @@ from unittest import TestCase
 from uuid import NAMESPACE_URL, UUID, uuid5
 
 from eventsourcing.application import (
+    AggregateNotFoundError,
     Application,
+    ProcessingEvent,
 )
 from eventsourcing.dcb.application import DCBApplication
 from eventsourcing.dcb.domain import EnduringObject, Tagged
 from eventsourcing.dcb.msgpack import Decision, InitialDecision
 from eventsourcing.dispatch import singledispatchmethod
-from eventsourcing.domain import Aggregate, event
+from eventsourcing.domain import Aggregate, DomainEventProtocol, event
 from eventsourcing.persistence import (
     IntegrityError,
     Tracking,
     TrackingRecorder,
 )
 from eventsourcing.projection import (
+    EventSourcedProjection,
+    EventSourcedProjectionRunner,
     Projection,
     ProjectionRunner,
 )
@@ -55,6 +59,32 @@ class EventCountersInterface(TrackingRecorder):
     @abstractmethod
     def incr_subsequent_event_counter(self, tracking: Tracking) -> None:
         pass
+
+
+class Counters(EventSourcedProjection[UUID]):
+    @singledispatchmethod
+    def policy(
+        self,
+        domain_event: DomainEventProtocol[UUID],
+        processing_event: ProcessingEvent[UUID],
+    ) -> None:
+        topic = get_topic(type(domain_event))
+        try:
+            counter_id = Counter.create_id(topic)
+            counter: Counter = self.repository.get(counter_id)
+        except AggregateNotFoundError:
+            counter = Counter(topic)
+        counter.increment()
+        processing_event.collect_events(counter)
+
+    def get_count(self, domain_event_class: type[DomainEventProtocol[UUID]]) -> int:
+        topic = get_topic(domain_event_class)
+        counter_id = Counter.create_id(topic)
+        try:
+            counter: Counter = self.repository.get(counter_id)
+        except AggregateNotFoundError:
+            return 0
+        return counter.count
 
 
 class EventCountersViewTestCase(TestCase):
@@ -353,3 +383,33 @@ class TaggedDecisionCountersProjectionTestCase(TestCase, ABC):
                     application_name=write_model.name,
                     notification_id=position,
                 )
+
+
+class EventSourcedProjectionTestCase(TestCase):
+    env: ClassVar[dict[str, str]] = {}
+
+    def test_event_sourced_projection(self) -> None:
+        with EventSourcedProjectionRunner(
+            application_class=Application, projection_class=Counters, env=self.env
+        ) as runner:
+            recordings = runner.app.save(Aggregate())
+            runner.wait(recordings[-1].notification.id)
+            self.assertEqual(1, runner.projection.get_count(Aggregate.Created))
+            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
+
+            recordings = runner.app.save(Aggregate())
+            runner.wait(recordings[-1].notification.id)
+            self.assertEqual(2, runner.projection.get_count(Aggregate.Created))
+            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
+
+            recordings = runner.app.save(Aggregate())
+            runner.wait(recordings[-1].notification.id)
+            self.assertEqual(3, runner.projection.get_count(Aggregate.Created))
+            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
+
+            aggregate = Aggregate()
+            aggregate.trigger_event(Aggregate.Event)
+            recordings = runner.app.save(aggregate)
+            runner.wait(recordings[-1].notification.id)
+            self.assertEqual(4, runner.projection.get_count(Aggregate.Created))
+            self.assertEqual(1, runner.projection.get_count(Aggregate.Event))
