+++ tests/projection_tests/test_event_sourced_projection_with_popo.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from typing import ClassVar
+
+from eventsourcing.tests.projection import EventSourcedProjectionTestCase
+
+
+class TestEventSourcedProjectionWithPOPO(EventSourcedProjectionTestCase):
+    env: ClassVar[dict[str, str]] = {"PERSISTENCE_MODULE": "eventsourcing.popo"}
+
+
+del EventSourcedProjectionTestCase
