+++ tests/projection_tests/test_event_sourced_projection_with_postgres.py
@@ -5,101 +5,20 @@ import multiprocessing.synchronize
 import traceback
 from threading import Thread
 from typing import TYPE_CHECKING, Any, ClassVar
-from unittest import TestCase
-from uuid import NAMESPACE_URL, UUID, uuid5
+from uuid import UUID
 
-from eventsourcing.application import (
-    AggregateNotFoundError,
-    Application,
-    ProcessingEvent,
-)
-from eventsourcing.dispatch import singledispatchmethod
-from eventsourcing.domain import Aggregate, DomainEventProtocol, event
+from eventsourcing.application import Application
+from eventsourcing.domain import Aggregate
 from eventsourcing.persistence import OperationalError
-from eventsourcing.projection import (
-    EventSourcedProjection,
-    EventSourcedProjectionRunner,
-)
+from eventsourcing.projection import EventSourcedProjectionRunner
 from eventsourcing.tests.postgres_utils import drop_tables, pg_close_all_connections
-from eventsourcing.utils import get_topic
+from eventsourcing.tests.projection import Counters, EventSourcedProjectionTestCase
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterable, Mapping
 
 
-class Counter(Aggregate):
-    def __init__(self, name: str) -> None:
-        self.name = name
-        self.count = 0
-
-    @classmethod
-    def create_id(cls, name: str) -> UUID:
-        return uuid5(NAMESPACE_URL, f"/counters/{name}")
-
-    @event("Incremented")
-    def increment(self) -> None:
-        self.count += 1
-
-
-class Counters(EventSourcedProjection[UUID]):
-    @singledispatchmethod
-    def policy(
-        self,
-        domain_event: DomainEventProtocol[UUID],
-        processing_event: ProcessingEvent[UUID],
-    ) -> None:
-        topic = get_topic(type(domain_event))
-        try:
-            counter_id = Counter.create_id(topic)
-            counter: Counter = self.repository.get(counter_id)
-        except AggregateNotFoundError:
-            counter = Counter(topic)
-        counter.increment()
-        processing_event.collect_events(counter)
-
-    def get_count(self, domain_event_class: type[DomainEventProtocol[UUID]]) -> int:
-        topic = get_topic(domain_event_class)
-        counter_id = Counter.create_id(topic)
-        try:
-            counter: Counter = self.repository.get(counter_id)
-        except AggregateNotFoundError:
-            return 0
-        return counter.count
-
-
-class TestEventSourcedProjection(TestCase):
-    def test_event_sourced_projection_popo(self) -> None:
-        with EventSourcedProjectionRunner(
-            application_class=Application,
-            projection_class=Counters,
-            env={
-                "PERSISTENCE_MODULE": "eventsourcing.popo",
-            },
-        ) as runner:
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(1, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(2, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(3, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            aggregate = Aggregate()
-            aggregate.trigger_event(Aggregate.Event)
-            recordings = runner.app.save(aggregate)
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(4, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(1, runner.projection.get_count(Aggregate.Event))
-
-
-class TestEventSourcedProjectionWithPostgres(TestCase):
+class TestEventSourcedProjectionWithPostgres(EventSourcedProjectionTestCase):
     env: ClassVar[dict[str, str]] = {
         "PERSISTENCE_MODULE": "eventsourcing.postgres",
         "POSTGRES_DBNAME": "eventsourcing",
@@ -114,34 +33,6 @@ class TestEventSourcedProjectionWithPostgres(TestCase):
 
     def tearDown(self) -> None:
         drop_tables()
-
-    def test_event_sourced_projection(self) -> None:
-        with EventSourcedProjectionRunner(
-            application_class=Application,
-            projection_class=Counters,
-            env=self.env,
-        ) as runner:
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(1, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(2, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            recordings = runner.app.save(Aggregate())
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(3, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(0, runner.projection.get_count(Aggregate.Event))
-
-            aggregate = Aggregate()
-            aggregate.trigger_event(Aggregate.Event)
-            recordings = runner.app.save(aggregate)
-            runner.wait(recordings[-1].notification.id)
-            self.assertEqual(4, runner.projection.get_count(Aggregate.Created))
-            self.assertEqual(1, runner.projection.get_count(Aggregate.Event))
 
     def test_server_closes_connections_before_run_forever(self) -> None:
         with EventSourcedProjectionRunner(
@@ -308,3 +199,6 @@ class TestEventSourcedProjectionWithPostgres(TestCase):
             self.assertIn("server closed the connection", str(exception))
             self.assertTrue(projection_errored.is_set())
             self.assertTrue(projection_stopped.is_set())
+
+
+del EventSourcedProjectionTestCase
